Automatically determine BUILD_X64 option (if unset).
Previously, this option was chosen based on the name of the currently
running schroot session. This would cause non-schroot-based builds to
fail unless the user manually passed something like "-DBUILD_X64=On".

Instead, we can examine the size of a void pointer to automatically
determine whether to do a 32-bit or 64-bit build. This should work for
both the schroot-based and non-schroot-based builds.

Note: If the user manually specifies -DBUILD_X64=On|Off at the
command-line, this still takes precedence over the automatic detection
of the native target.