Move top-level CMakeLists.txt file to the actual top-level.
It's a bit unorthodox to have to do "cmake src" rather than "cmake .",
(or similarly, within a build directory, to have to do "cmake ../src"
rather than "cmake .."). So move this file to the top level to support
the more orthodox form.

Additionally, the presence of the top-level CMakeLists.txt file
clearly advertises that this project is using cmake, so the alert
reader can easily guess how to build things.