vogleditor: Add for additional nest operations
Add glPush/PopDebugGroup as nesting delimiters along with existing
glBegin/End. Use same method as for other "vogl_is_xxx_entrypoint"
operations.

Comments:
If this glPush/PopDebugGroup addition is (and/or possibly others are)
accepted for nesting, another method would be to add a flag attribute
to packets to identify apicall types that can be logically AND'd rather
than using a switch statement (as done in apitrace).

Alternatively, create a "nest" object to encapsulate this information,
initializing it with a set of nested pairs (statically or via some
preference) and provide an interface(s) to return needed information.

Signed-off-by: Lawrence L Love <lawrencex.l.love@intel.com>