Initial ktxtool & pxfmt support for S3TC/DXT compressed textures.
There's a new pxfmt_decompress_pixels() function to decompress 1 pixel at a
time (similar to the "fetch" functions within Mesa3D).  The pxfmt library is
leveraging the following parts of Mesa3D:

- src/mesa/main/texcompress_s3tc.c
- src/mesa/main/dlopen.h

It was modified for the following reasons:

1) To minimize how much of Mesa3D needs to be imported (there were lots of
   dependendcies to the rest of Mesa3D, when trying to use files "as-is").

2) To focus on decompression (compression isn't needed).

3) To fit within the pxfmt library.  This was more of a tactic to accomplish
   the 1st reason.

The fetch functions from "texcompress_s3tc.c" were abstracted and ported.
Since this code uses an external library (because of legal concerns), the
"dlopen.h" file was also ported.

NOTE: I had to work-around a problem for RGBA_DXT1.  The OpenGL extension
specification for DXT1 says that each block is 64 bits, but the VOGL KTX files
use 128-bits-per-block for RGBA_DXT1.

Since the external library is provided by default on some Linux distros, and
not on Windows, there is no DXT support on Windows (yet).