Workaround to skip binding missing RenderBuffers rather than asserting
For a Dota2 trace there seems to be a snapshot bug that is not capturing all of the Renderbuffer state associated with FBs.  Upon restore I'm hitting an error due to various RB handles being invalid.  This change works around that error for the time being until root cause is fixed.