First working version of --fs_preprocessor option
Use "--fs_preprocessor <path>" to point to tool that all Frag Shaders will be sent to prior to glShaderSource() call.  Output of tool will be used as alternate input to glShaderSource().

Use "--fs_preprocessor_options <options>" to specify options that will be passed to the tool given above.

Current Limitations:
-Only FSs that are captured in an api trace glShaderSource() call will run through the pre-processor. Shaders restored from a snapshot at the start of the a trimmed trace will not go through pre-processor.
-Right now tmp in/out frag shaders are stored in 2 files named tmp_shader_[in|out].frag in vogl exe dir which files are not currently deleted.