UI: Add State/Render groups option
This addresses two items from issue# 161.

> 1) Add a way to toggle between modes so that users can easily just
> see the draw calls without other clutter and clicking - perhaps a
> combo box so that we can easily add more hierarchy methods without
> consuming too much toolbar space?

The combo box is in development. But, if having the option of groups
is considered advantageous, it seems like it would be a good idea to
get the infrastructure in now since there is other work going on with
vogleditor. It would be easier for future vogleditor-related commits
to branch on this than merge later.

I've added an option to the settings dialog to enable/disable groups
in the tree display. The default is to display the tree as previously.

> 2) Another request is to support glPushDebugGroup/glPopDebugGroup
> separately from the state changes. I'd expect developers to know what
> their debug group names mean, but just seeing "State Changes" may not
> tell them anything useful

Debug markers (glPushDebugGroup/glPopDebugGroup) calls are now
disengaged from State/Render groups. If a Debug marker is initiated
then, if Render/State groups are enabled, any State/Render group
collection ends until the Debug marker collection ends.

Development can continue with the other items mentioned in issue# 161
without affecting the default apicall tree display.

Additional details
------------------
The current settings are static and only apply at session startup.
The user should see no difference in the tree display unless
State/Render groups are enabled for the next session.

Other calls can easily be added for nesting (e.g., glPush/PopMatrix,
et al). The plan is to control these along with the other grouping
options dynamically from the settings dialog combo box.

In vogl_gl_utils.cpp a vogl_is_frame_buffer_write function was added
that includes vogl_is_draw_entrypoint, vogl_is_clear_entrypoint, glEnd
and glBitmap. This function is used for Render groups but also replaced
code in vogleditor_qapicalltreemodel.cpp that was checking for all four
in single if-blocks, e.g.,

    if (vogl_is_draw_entrypoint(entrypointId) ||
        vogl_is_clear_entrypoint(entrypointId) ||
        (entrypointId == VOGL_ENTRYPOINT_glBitmap) ||
        (entrypointId == VOGL_ENTRYPOINT_glEnd))

Testing
-------
I have a number of .bin files I used that I can share

    debug_khr_matrix.bin
    glxgears-pushpop.bin
    glxgears-unpaired-glend.bin
    glxspheres64.bin
    mydebug_khr-xtra-pop.bin
    mydebug_khr-xtra-push.bin
    nested-glBegin.bin

Export API Calls exports exactly what you see in the tree but once
there is dynamic control then it will be easy to switch to the another
view, if desired, before exporting.