UI: move some logic to constructors
Some code cleanup while reading through the code:

In creating the apicall tree, the client (vogleditor_qapicalltreemodel)
creates the calltree objects (vogleditor_apiCallTreeItem objects). If
the object is constructed with a parent then the client must then add
the new object to the child list of the object's parent after the object
has been constructed.

Instead of the client having to know/remember to do this for the object,
it's now done in the object's constructor. This makes for easier
comprehension/reading of the code in my opinion.

Signed-off-by: Lawrence L Love <lawrencex.l.love@intel.com>