UI: Add "Elapsed" column for State/Render groups
The "Duration" column contents of a State/Render group was the
difference between the start time of its first apicall and end
time of its last apicall. But on inspection, this duration time
did not always add up to the sum of its constituent children's
duration times.

This patch changes the calculation of the duration time for a
State/Render group to be the aggregate of the duration times of
its children.

In addition, an "Elapsed" column was added which only appears when the
"State/Render groups" setting is selected and is only populated for a
State/Render group. This column contains the calculation of the original
State/Render group duration time (i.e., start time of first apicall to
end time of last apicall).

If useful, duration and elapsed times could also be implemented for
frames.