UI: remove multiple rand() calls in randomRGB()
Initially, this commit was to modify the randomRGB() function to return
values along the blue and green or blue and red spectrums to distinguish
the colors of Debug marker groups from from individual API calls which
are colored along the green and red spectrum.

But since the current default is to only color Debug marker groups if
State/Render groups are also selected and which is enforced by recent
commit:

   UI: Color-coding Debug groups shouldn't be default

then individual API calls won't be displayed. With both State/Render
groups and Debug marker groups selected, all API calls are contained
by either a State, Draw, or Debug marker group.

If it is decided at some point that the default should be to always
color Debug marker groups independent of State/Render groups then there
are some options to consider:

1) I've left in commented code to return blue/red and blue/green colors
   as described above to distinguish these groups from red/green colored
   API calls.

2) If more distinguishable colors are desired for Debug marker groups
   than given from the commented algorithm:

   Use a predefined table of colors (many defined online); cycle thru
   them in a % modulus fashion. This assumes it's OK to repeat colors
   if the number of groups exceeds the number of stored colors.

3) If we do not need individual and distingusiable colors for Debug
   marker groups:

   Just alternate between two colors in the blue/green and blue/red
   spectrums (e.g., cyan and magenta) with no need to generate colors.

Signed-off-by: Lawrence L Love <lawrencex.l.love@intel.com>