Fix for Issue #148 The API tree view now displays human readable values for GLbitfields e.g. glClear( GL_DEPTH_BUFFER_BIT | GL_COLOR_BUFFER_BIT ).
Changes:
Modified pretty_print_param_val() to be able to output pretty interpretations of GLbitfields.
If it finds that the parameter is of type VOGL_GLBITFIELD then it will test each bit in turn and if it is set will look for an appropriate name for that mask.
It separates each mask with a "|", displays a decimal representation if no name is found for a mask and simply displays "0" if no bits are set.