Items on the time line can now be clicked to display the matching API call in the call tree view. Hovering over an item on the time line now displays the call number, name and parameters in a tool tip.
Changes:
New function vogleditor_QTimelineView::itemUnderPos determines the vogleditor_timelineItem* that is under the given position with a 1 pixel tolerance.
This is used by vogleditor_QTimelineView::mouseReleaseEvent for clicks and vogleditor_QTimelineView::event for displaying tool tips. Tool tips are displayed when QEvent::ToolTip is received.
In order to select an item from the tree view a new slot VoglEditor::selectAPICallItem was needed which is connected to the new signal QTimelineView::timelineItemClicked.
VoglEditor::selectAPICallItem uses QAbstractItemModel::match to find the QModelIndex for that call, this now requires Qt::UserRole to be set to the vogleditor_apiCallItem* for that item. This is an implementation of option 1 described here: https://wiki.qt.io/Qt_project_org_faq#How_can_a_QModelIndex_be_retrived_from_the_model_for_an_internal_data_item.3F